Remove split from selection from this PR
